Run v7 generator (#31)
* Run v7 generator

* symlink test->tests for hatch compatibility

generator writes to test directory, hatch configuration looks for tests
and it's messy trying to force `hatch test` to use the test directory
without having to call `hatch run` with a custom config.

* Support explcit version during generation

run generate script with version 3.0.0

Bumping to major version due to significant changes to the generated clients.